General update

+++ src/main/java/com/application/drugsProject/repository/DrugRepository.java
@@ -3,7 +3,7 @@ package com.application.drugsProject.repository;
 import com.application.drugsProject.model.Drug;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
 @Repository
-public interface DrugRepository extends JpaRepository<Drug, Integer> {
+public interface DrugRepository extends JpaRepository<Drug, Long> {
+
 }
